pyproject.toml
[tool.poetry]
name = "videosdb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.1.4"
Autologging = "^1.3.2"
python-wordpress-xmlrpc = "^2.3"
youtube-transcript-api = "^0.3.1"
Jinja2 = "^2.11.2"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"
django-uuslug = "^1.2.0"
psycopg2-binary = "^2.8.6"
django-dirtyfields = "^1.4.1"
google-api-python-client = "^1.12.8"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.4"
pylint = "^2.6.0"
django-extensions = "^3.1.0"
pip-autoremove = "^0.9.1"
autopep8 = "^1.5.4"
jedi = "<0.18"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
